bổ sung check phân quyền cho mỗi api

+++ CodeBE_COMP1640/CodeBE_COMP1640/Controllers/ArticleController/ArticleController.cs
@@ -1,5 +1,7 @@
-﻿using CodeBE_COMP1640.Factories.Implements;
+﻿using CodeBE_COMP1640.Controllers.PermissionController;
+using CodeBE_COMP1640.Factories.Implements;
 using CodeBE_COMP1640.Services.EmailS;
+using CodeBE_COMP1640.Services.PermissionS;
 using CodeBE_COMP1640.Services.UserS;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -18,10 +20,11 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         private readonly IEmailSender _emailSender;
 
         private readonly IUserService _userService;
+        private readonly IPermissionService PermissionService;
         private readonly JsonSerializerOptions _jsonSerializerOptions;
         private static readonly Dictionary<string, List<(string originalFileName, byte[] fileBytes)>> _articleFiles = new Dictionary<string, List<(string originalFileName, byte[] fileBytes)>>();
 
-        public ArticleController(IServiceProvider serviceProvider, IConfiguration configuration, IEmailSender emailSender, IUserService userService)
+        public ArticleController(IServiceProvider serviceProvider, IConfiguration configuration, IEmailSender emailSender, IUserService userService, IPermissionService permissionService)
         {
             _serviceProvider = serviceProvider;
             _configuration = configuration;
@@ -33,13 +36,19 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
                 ReferenceHandler = ReferenceHandler.Preserve,
                 MaxDepth = 16
             };
+            PermissionService = permissionService;
         }
 
         [Route(ArticleRoute.Get), HttpGet, Authorize]
-        public IActionResult Get(int id)
+        public async Task<IActionResult> Get(int id)
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 var data = _repositoryFactory.ArticleRepository.Get(id);
                 return Ok(new
                 {
@@ -60,6 +69,11 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 // Tạo entity từ request và đặt IsApproved là false
                 var articleEntity = request.ToEntity();
                 articleEntity.IsApproved = false;
@@ -109,10 +123,15 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         }
 
         [Route(ArticleRoute.Update), HttpPut, Authorize]
-        public IActionResult Update(int id, ArticlePut request)
+        public async Task<IActionResult> Update(int id, ArticlePut request)
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 var entity = request.ToEntity();
                 entity.ArticleId = id;
                 var data = _repositoryFactory.ArticleRepository.Update(request.ToEntity());
@@ -131,10 +150,15 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         }
 
         [Route(ArticleRoute.Delete), HttpDelete, Authorize]
-        public IActionResult Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 var data = _repositoryFactory.ArticleRepository.Delete(id);
                 return Ok(new
                 {
@@ -151,10 +175,15 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         }
 
         [Route(ArticleRoute.ListArticle), HttpGet, Authorize]
-        public IActionResult GetAllArticles()
+        public async Task<IActionResult> GetAllArticles()
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 var allArticles = _repositoryFactory.ArticleRepository.GetAll();
                 return Ok(new
                 {
@@ -171,10 +200,15 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         }
 
         [Route(ArticleRoute.GetByUser), HttpGet, Authorize]
-        public IActionResult GetByUser(int userId)
+        public async Task<IActionResult> GetByUser(int userId)
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 var data = _repositoryFactory.ArticleRepository.GetListByUserId(userId);
                 return Ok(new
                 {
@@ -191,10 +225,15 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         }
 
         [Route(ArticleRoute.GetByDepartment), HttpGet, Authorize]
-        public IActionResult GetByDepartment(int departmentId)
+        public async Task<IActionResult> GetByDepartment(int departmentId)
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 var data = _repositoryFactory.ArticleRepository.GetListByDepartmentId(departmentId);
                 return Ok(new
                 {
@@ -211,10 +250,15 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         }
 
         [Route(ArticleRoute.Approve), HttpPut, Authorize]
-        public IActionResult ApproveArticle(int articleId)
+        public async Task<IActionResult> ApproveArticle(int articleId)
         {
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 // Lấy article từ repository
                 var article = _repositoryFactory.ArticleRepository.Get(articleId);
 
@@ -250,6 +294,11 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
 
             try
             {
+                if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+                {
+                    return Forbid();
+                }
+
                 if (string.IsNullOrEmpty(articleId))
                     return BadRequest("Invalid articleId");
 
@@ -301,8 +350,13 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
         }
 
         [Route(ArticleRoute.GetFile), HttpGet, Authorize]
-        public IActionResult GetFile(string articleId)
+        public async Task<IActionResult> GetFile(string articleId)
         {
+            if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+            {
+                return Forbid();
+            }
+
             if (_articleFiles.ContainsKey(articleId))
             {
                 var fileList = _articleFiles[articleId];
@@ -334,8 +388,13 @@ namespace CodeBE_COMP1640.Controllers.ArticleController
 
 
         [Route(ArticleRoute.Export), HttpGet, Authorize]
-        public IActionResult ExportFiles()
+        public async Task<IActionResult> ExportFiles()
         {
+            if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+            {
+                return Forbid();
+            }
+
             string uploadFolder = "./UploadFile";
 
             if (!Directory.Exists(uploadFolder) || !Directory.EnumerateFiles(uploadFolder).Any())
