+++ CodeBE_COMP1640/CodeBE_COMP1640/Controllers/FeedbackController/FeedbackController.cs
@@ -10,7 +10,8 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using CodeBE_COMP1640.Services.EmailS;
-
+using CodeBE_COMP1640.Services.PermissionS;
+using CodeBE_COMP1640.Controllers.PermissionController;
 
 namespace CodeBE_COMP1640.Controllers.FeedbackController
 {
@@ -23,11 +24,13 @@ namespace CodeBE_COMP1640.Controllers.FeedbackController
 
         private readonly IUserService _userService;
 
-        public FeedbackController(IFeedbackService feedbackService, IEmailSender emailSender, IUserService userService)
+        private readonly IPermissionService PermissionService;
+
+        public FeedbackController(IFeedbackService feedbackService, IEmailSender emailSender, IUserService userService, IPermissionService PermissionService)
         {
             _feedbackService = feedbackService ?? throw new ArgumentNullException(nameof(feedbackService));
-
             this._emailSender = emailSender;
+            this.PermissionService = PermissionService;
             _userService = userService ?? throw new ArgumentNullException(nameof(userService));
 
         }
@@ -35,6 +38,11 @@ namespace CodeBE_COMP1640.Controllers.FeedbackController
         [Route(FeedbackRoute.List), HttpGet, Authorize]
         public async Task<ActionResult<List<Feedback>>> GetAllFeedbacks()
         {
+            if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+            {
+                return Forbid();
+            }
+
             var feedbacks = await _feedbackService.GetFeedbacks();
             return Ok(feedbacks);
         }
@@ -42,6 +50,11 @@ namespace CodeBE_COMP1640.Controllers.FeedbackController
         [Route(FeedbackRoute.Get), HttpGet, Authorize]
         public async Task<ActionResult<Feedback>> GetFeedbackById(int id)
         {
+            if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+            {
+                return Forbid();
+            }
+
             var feedback = await _feedbackService.GetFeedbackById(id);
             if (feedback == null)
             {
@@ -52,39 +65,38 @@ namespace CodeBE_COMP1640.Controllers.FeedbackController
 
         [Route(FeedbackRoute.Create), HttpPost, Authorize]
         public async Task<IActionResult> CreateFeedback(FeedbackDTO feedbackDTO)
-{               
-               
-                
-                // Tạo mới đối tượng Feedback từ dữ liệu trong request body
-                var feedback = new Feedback
-                
-                {
-                    UserId = feedbackDTO.UserId,
-                    ArticleId = feedbackDTO.ArticleId,
-                    FeedbackContent = feedbackDTO.FeedbackContent,
-                    FeedbackTime = feedbackDTO.FeedbackTime,
-                    
-                };
-                // Gọi phương thức tạo mới feedback từ service
-                await _feedbackService.CreateFeedback(feedback);
-                 // Lấy thông tin người dùng từ UserService
-                var user = await _userService.Get(feedbackDTO.UserId);
-                
-                //gửi email
-                 if (user != null)
-                    {   
-                        var receivers = new List<string>();
-                        receivers.Add(user.Email);
-                        var subject = "Test";
-                        var message = "Hello";
-                        await _emailSender.SendEmailAsync(receivers, subject, message);
-                    }
-            
-                return CreatedAtAction(nameof(GetFeedbackById), new { id = feedback.FeedbackId }, feedback);
-                
-        
+        {
+            if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+            {
+                return Forbid();
+            }
 
-          
+            // Tạo mới đối tượng Feedback từ dữ liệu trong request body
+            var feedback = new Feedback
+
+            {
+                UserId = feedbackDTO.UserId,
+                ArticleId = feedbackDTO.ArticleId,
+                FeedbackContent = feedbackDTO.FeedbackContent,
+                FeedbackTime = feedbackDTO.FeedbackTime,
+
+            };
+            // Gọi phương thức tạo mới feedback từ service
+            await _feedbackService.CreateFeedback(feedback);
+            // Lấy thông tin người dùng từ UserService
+            var user = await _userService.Get(feedbackDTO.UserId);
+
+            //gửi email
+            if (user != null)
+            {
+                var receivers = new List<string>();
+                receivers.Add(user.Email);
+                var subject = "Test";
+                var message = "Hello";
+                await _emailSender.SendEmailAsync(receivers, subject, message);
+            }
+
+            return CreatedAtAction(nameof(GetFeedbackById), new { id = feedback.FeedbackId }, feedback);
         }
 
 
@@ -92,6 +104,11 @@ namespace CodeBE_COMP1640.Controllers.FeedbackController
         [Route(FeedbackRoute.Update), HttpPut, Authorize]
         public async Task<IActionResult> UpdateFeedback(int id, FeedbackDTO feedbackDTO)
         {
+            if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+            {
+                return Forbid();
+            }
+
             var existingFeedback = await _feedbackService.GetFeedbackById(id);
             if (existingFeedback == null)
             {
@@ -111,6 +128,11 @@ namespace CodeBE_COMP1640.Controllers.FeedbackController
         [Route(FeedbackRoute.Delete), HttpDelete, Authorize]
         public async Task<IActionResult> DeleteFeedback(int id)
         {
+            if (!await PermissionService.HasPermission(PermissionRoute.ListPermission, PermissionService.GetUserId()))
+            {
+                return Forbid();
+            }
+
             await _feedbackService.DeleteFeedback(id);
             return NoContent();
         }
